courier-service/backend/requirements.txt
Django>=5.0,<6.0
PyMySQL>=1.1,<2.0
python-dotenv>=1.0,<2.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
